Adopt ERC20 plugin, move registry to integrations; dissolve the folder
- ERC20 now comes from @owneraio/finp2p-ethereum-erc20-plugin@^0.28.1
  (drop-in: zero-arg ctor, custody-wallet mint/deploy — same semantics);
  built-in erc20.ts deleted.
- Registry (TokenStandardRegistry + singleton) moved to
  integrations/token-standards/registry.ts, co-located with the registrar;
  all consumers repointed.
- services/direct/token-standards/ folder removed entirely (shims already
  dropped, erc20 to plugin, registry relocated).
