+++ src/integrations/token-standards/index.ts
@@ -1,1 +1,2 @@
 export * from "./register";
+export * from "./registry";
