+++ src/services/direct/index.ts
@@ -9,7 +9,6 @@ export * from './omnibus-delegate'
 export * from './mapping-validator'
 export * from './mapping-config'
 export * from './wallet-resolver'
-export * from './token-standards'
 export * from './local-submit-signer'
 export * from './gas-prefunding-option'
 export * from './token-whitelisting-option'
