--- src/services/direct/token-standards/index.ts
@@ -1,2 +1,0 @@
-export * from './registry';
-export * from './erc20';
